Move lst_from_uv to uvflag to avoid circular imports
